extract tod and shiftitemcontent
